# src/adapters/rest/generated/src/example.adapters.rest.generated/Controllers/AnswerQuestionTasksApi.cs
/*
 * Contact center service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using example.adapters.rest.generated.Attributes;
using example.adapters.rest.generated.Models;

namespace example.adapters.rest.generated.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class AnswerQuestionTasksApiController : ControllerBase
    { 
        /// <summary>
        /// Completes the answer question task by answering the question.
        /// </summary>
        /// <param name="taskId"></param>
        /// <param name="answerQuestion"></param>
        /// <response code="202">Accepted</response>
        /// <response code="404">Task not found</response>
        [HttpPut]
        [Route("/api/questions/answer_question_tasks/{task_id}/answer")]
        [Consumes("application/json")]
        [ValidateModelState]
        public abstract Task<IActionResult> AnswerQuestion([FromRoute (Name = "task_id")][Required]long taskId, [FromBody]AnswerQuestion answerQuestion);

        /// <summary>
        /// Gets an answer question task.
        /// </summary>
        /// <param name="taskId"></param>
        /// <response code="200">OK</response>
        /// <response code="404">Task not found</response>
        [HttpGet]
        [Route("/api/questions/answer_question_tasks/{task_id}")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(AnswerQuestionTask))]
        public abstract Task<IActionResult> GetAnswerQuestionTask([FromRoute (Name = "task_id")][Required]long taskId);
    }
}