# src/adapters/rest/generated/src/example.adapters.rest.generated/Controllers/ModifyAnswerTasksApi.cs
/*
 * Contact center service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using example.adapters.rest.generated.Attributes;
using example.adapters.rest.generated.Models;

namespace example.adapters.rest.generated.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class ModifyAnswerTasksApiController : ControllerBase
    { 
        /// <summary>
        /// Gets a modify answer task.
        /// </summary>
        /// <param name="taskId"></param>
        /// <response code="200">OK</response>
        /// <response code="404">Task not found</response>
        [HttpGet]
        [Route("/api/questions/modify_answer_tasks/{task_id}")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(ModifyAnswerTask))]
        public abstract Task<IActionResult> GetModifyAnswerTask([FromRoute (Name = "task_id")][Required]long taskId);

        /// <summary>
        /// Completes the modify answer task by modifing the answer.
        /// </summary>
        /// <param name="taskId"></param>
        /// <param name="modifyAnswer"></param>
        /// <response code="202">Accepted</response>
        /// <response code="404">Task not found</response>
        [HttpPut]
        [Route("/api/questions/modify_answer_tasks/{task_id}/modify")]
        [Consumes("application/json")]
        [ValidateModelState]
        public abstract Task<IActionResult> ModifyAnswer([FromRoute (Name = "task_id")][Required]long taskId, [FromBody]ModifyAnswer modifyAnswer);
    }
}